# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MBG Dapur - Sistem Manajemen Operasional Dapur
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /menu:
    get:
      operationId: listMenu
      tags: [menu]
      summary: List menu harian
      parameters:
        - name: tanggal
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List menu
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Menu"
    post:
      operationId: createMenu
      tags: [menu]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Menu"
  /menu/{id}:
    get:
      operationId: getMenu
      tags: [menu]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Menu"
    patch:
      operationId: updateMenu
      tags: [menu]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Menu"
    delete:
      operationId: deleteMenu
      tags: [menu]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # PRODUKSI
...
  /pengiriman:
    get:
      operationId: listPengiriman
      tags: [pengiriman]
      summary: List manifest pengiriman
      parameters:
        - name: tanggal
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: driver_id
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List pengiriman
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pengiriman"
    post:
      operationId: createPengiriman
      tags: [pengiriman]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PengirimanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pengiriman"
  /pengiriman/{id}:
    get:
      operationId: getPengiriman
      tags: [pengiriman]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pengiriman detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pengiriman"
    patch:
      operationId: updatePengiriman
      tags: [pengiriman]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PengirimanUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pengiriman"
...
  /penerima-manfaat:
    get:
      operationId: listPenerimaManfaat
      tags: [penerima-manfaat]
      summary: List penerima manfaat
      parameters:
        - name: sekolah
          in: query
          schema:
            type: string
        - name: wilayah
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List penerima manfaat
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PenerimaManfaat"
    post:
      operationId: createPenerimaManfaat
      tags: [penerima-manfaat]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PenerimaManfaatInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PenerimaManfaat"
  /penerima-manfaat/{id}:
    patch:
      operationId: updatePenerimaManfaat
      tags: [penerima-manfaat]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PenerimaManfaatUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PenerimaManfaat"
    delete:
      operationId: deletePenerimaManfaat
      tags: [penerima-manfaat]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /verifikasi-penerimaan:
    post:
      operationId: createVerifikasiPenerimaan
      tags: [penerima-manfaat]
      summary: Verifikasi penerimaan makanan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifikasiPenerimaanInput"
      responses:
        "201":
          description: Verifikasi dibuat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifikasiPenerimaan"
...
  /keuangan/anggaran:
    get:
      operationId: listAnggaran
      tags: [keuangan]
      summary: List anggaran per dapur
      parameters:
        - name: dapur_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: periode
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List anggaran
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Anggaran"
    post:
      operationId: createAnggaran
      tags: [keuangan]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnggaranInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Anggaran"
  /keuangan/anggaran/{id}:
    patch:
      operationId: updateAnggaran
      tags: [keuangan]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnggaranUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Anggaran"
  /keuangan/realisasi:
    get:
      operationId: listRealisasi
      tags: [keuangan]
      summary: List realisasi biaya
      responses:
        "200":
          description: List realisasi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Realisasi"
    post:
      operationId: createRealisasi
      tags: [keuangan]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RealisasiInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Realisasi"
...
components:
  schemas:
...
    Menu:
      type: object
      required: [id, nama, tanggal, kategori, target_porsi, created_at]
      properties:
        id:
          type: integer
        nama:
          type: string
        deskripsi:
          type: ["string", "null"]
        tanggal:
          type: string
        kategori:
          type: string
        target_porsi:
          type: integer
        kalori:
          type: ["number", "null"]
        created_at:
          type: string
    MenuInput:
      type: object
      required: [nama, tanggal, kategori, target_porsi]
      properties:
        nama:
          type: string
        deskripsi:
          type: string
        tanggal:
          type: string
        kategori:
          type: string
        target_porsi:
          type: integer
        kalori:
          type: number
    MenuUpdate:
      type: object
      properties:
        nama:
          type: string
        deskripsi:
          type: string
        tanggal:
          type: string
        kategori:
          type: string
        target_porsi:
          type: integer
        kalori:
          type: number

    # PRODUKSI
...
    Pengiriman:
      type: object
      required: [id, dapur_id, tanggal, status, created_at]
      properties:
        id:
          type: integer
        dapur_id:
          type: integer
        dapur_nama:
          type: ["string", "null"]
        driver_id:
          type: ["integer", "null"]
        driver_nama:
          type: ["string", "null"]
        tanggal:
          type: string
        jumlah_porsi:
          type: integer
        tujuan:
          type: string
        status:
          type: string
          enum: [dijadwalkan, berangkat, tiba, selesai, gagal]
        catatan:
          type: ["string", "null"]
        created_at:
          type: string
    PengirimanInput:
      type: object
      required: [dapur_id, tanggal, jumlah_porsi, tujuan]
      properties:
        dapur_id:
          type: integer
        driver_id:
          type: ["integer", "null"]
        tanggal:
          type: string
        jumlah_porsi:
          type: integer
        tujuan:
          type: string
        catatan:
          type: string
    PengirimanUpdate:
      type: object
      properties:
        driver_id:
          type: ["integer", "null"]
        status:
          type: string
        catatan:
          type: string
        jumlah_porsi:
          type: integer
...
    PenerimaManfaat:
      type: object
      required: [id, nama, sekolah, kelas, wilayah, created_at]
      properties:
        id:
          type: integer
        nama:
          type: string
        sekolah:
          type: string
        kelas:
          type: string
        wilayah:
          type: string
        is_aktif:
          type: boolean
        created_at:
          type: string
    PenerimaManfaatInput:
      type: object
      required: [nama, sekolah, kelas, wilayah]
      properties:
        nama:
          type: string
        sekolah:
          type: string
        kelas:
          type: string
        wilayah:
          type: string
    PenerimaManfaatUpdate:
      type: object
      properties:
        nama:
          type: string
        sekolah:
          type: string
        kelas:
          type: string
        wilayah:
          type: string
        is_aktif:
          type: boolean
    VerifikasiPenerimaanInput:
      type: object
      required: [penerima_manfaat_id, tanggal, status]
      properties:
        penerima_manfaat_id:
          type: integer
        tanggal:
          type: string
        status:
          type: string
          enum: [hadir, tidak_hadir]
        catatan:
          type: string
    VerifikasiPenerimaan:
      type: object
      required: [id, penerima_manfaat_id, tanggal, status, created_at]
      properties:
        id:
          type: integer
        penerima_manfaat_id:
          type: integer
        penerima_nama:
          type: ["string", "null"]
        tanggal:
          type: string
        status:
          type: string
        catatan:
          type: ["string", "null"]
        created_at:
          type: string
...
    Anggaran:
      type: object
      required: [id, dapur_id, periode, total_anggaran, created_at]
      properties:
        id:
          type: integer
        dapur_id:
          type: integer
        dapur_nama:
          type: ["string", "null"]
        periode:
          type: string
        total_anggaran:
          type: number
        anggaran_per_porsi:
          type: ["number", "null"]
        created_at:
          type: string
    AnggaranInput:
      type: object
      required: [dapur_id, periode, total_anggaran]
      properties:
        dapur_id:
          type: integer
        periode:
          type: string
        total_anggaran:
          type: number
        anggaran_per_porsi:
          type: number
    AnggaranUpdate:
      type: object
      properties:
        total_anggaran:
          type: number
        anggaran_per_porsi:
          type: number
    Realisasi:
      type: object
      required: [id, dapur_id, tanggal, kategori, jumlah, created_at]
      properties:
        id:
          type: integer
        dapur_id:
          type: integer
        dapur_nama:
          type: ["string", "null"]
        tanggal:
          type: string
        kategori:
          type: string
          enum: [bahan_baku, operasional, sdm, sewa, lainnya]
        jumlah:
          type: number
        deskripsi:
          type: ["string", "null"]
        created_at:
          type: string
    RealisasiInput:
      type: object
      required: [dapur_id, tanggal, kategori, jumlah]
      properties:
        dapur_id:
          type: integer
        tanggal:
          type: string
        kategori:
          type: string
        jumlah:
          type: number
        deskripsi:
          type: string
...