# lib/api-zod/src/generated/types/anggaranUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MBG Dapur - Sistem Manajemen Operasional Dapur
 * OpenAPI spec version: 0.1.0
 */

export interface AnggaranUpdate {
  total_anggaran?: number;
  anggaran_per_porsi?: number;
}