# lib/api-zod/src/generated/types/pengiriman.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MBG Dapur - Sistem Manajemen Operasional Dapur
 * OpenAPI spec version: 0.1.0
 */
import type { PengirimanStatus } from './pengirimanStatus';

export interface Pengiriman {
  id: number;
  dapur_id: number;
  /** @nullable */
  dapur_nama?: string | null;
  /** @nullable */
  driver_id?: number | null;
  /** @nullable */
  driver_nama?: string | null;
  tanggal: string;
  jumlah_porsi?: number;
  tujuan?: string;
  status: PengirimanStatus;
  /** @nullable */
  catatan?: string | null;
  created_at: string;
}